added some code in backend part

+++ Mern_app/backend/controllers/product.controller.js
@@ -30,6 +30,11 @@ export const createProduct = async (req, res) => {
 
 export const DeleteProduct = async (req, res) => {
   const { id } = req.params;
+  
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({ success: false, msg: "inavalid product id" });
+  }
+  
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, msg: "Product deleted" });
